refactor(app.js): Extract status pages
